Implement feedback for file upload tests
RISDEV-175
